Add translation support for most admin pages.
git-svn-id: https://simplesamlphp.googlecode.com/svn/trunk@711 44740490-163a-0410-bde0-09ae8108e29a

+++ templates/default/admin-metadatalist.php
@@ -1,15 +1,30 @@
-<?php $this->includeAtTemplateBase('includes/header.php'); ?>
+<?php
+$this->data['header'] = $this->t('metaover_header');
+$this->data['icon'] = 'bino.png';
+
+$this->includeAtTemplateBase('includes/header.php');
+?>
 
 	
 	<div id="content">
 
-		<p>To look at the details for an SAML entity, click on the SAML entity header.</p>
+		<p><?php echo $this->t('metaover_intro'); ?></p>
 
 		<?php
 		
 		
-		function showEntry($header, $list, $baseurl) {
-		
+		function showEntry($t, $id) {
+
+			if (!array_key_exists($id, $t->data)) {
+				/* This metadata does not exist. */
+				return;
+			}
+
+
+			$header = $t->t('metaover_group_' . $id);
+			$list = $t->data[$id];
+			$baseurl = $t->data['baseurlpath'];
+
 			echo '<h3>' . $header . '</h3>';
 			
 
@@ -32,7 +47,7 @@
 				
 				if ($warning) {
 					echo '<div><img src="/' . $baseurl . 'resources/icons/caution.png" style="float: left; margin-right: 1em" />';
-					echo 'Error in this metadata entry.</div>';
+					echo $t->t('metaover_errorentry') . '</div>';
 				}
 				
 				echo '<div id="metadatasection-' . $encodedEntityID . '" style="display: none">';
@@ -42,7 +57,7 @@
 				}
 				
 				echo '<div style="margin-left: 1em">';
-				echo '<div class="efieldlist"><h5>Required fields</h5>';
+				echo '<div class="efieldlist"><h5>' . $t->t('metaover_required') . '</h5>';
 				echo '<dl>';
 				foreach ($entity['required.found'] AS $key => $value) {
 					echo '<dt>' . htmlspecialchars($key) . '</dt><dd>' . htmlspecialchars(var_export($value, TRUE)) . '</dd>';
@@ -53,7 +68,7 @@
 	
 				if (count($entity['required.notfound']) > 0) {
 					echo '</div><div class="efieldlist warning">';
-					echo '<h5>The following required fields was not found</h5><ul>';
+					echo '<h5>' . $t->t('metaover_required_not_found') . '</h5><ul>';
 					foreach ($entity['required.notfound'] AS $key) {
 						echo '<li>' . htmlspecialchars($key) . '</li>';
 					}
@@ -63,7 +78,7 @@
 				
 				if (count($entity['optional.found']) > 0) {
 					echo '</div><div class="efieldlist">';
-					echo '<h5>Optional fields</h5>';
+					echo '<h5>' . $t->t('metaover_optional_found') . '</h5>';
 					echo '<dl>';
 					foreach ($entity['optional.found'] AS $key => $value) {
 						echo '<dt>' . htmlspecialchars($key) . '</dt><dd>' . htmlspecialchars(var_export($value, TRUE)) . '</dd>';
@@ -75,7 +90,7 @@
 	
 				if (count($entity['optional.notfound']) > 0) {
 					echo '</div><div class="efieldlist info">';				
-					echo '<h5>The following optional fields was not found:</h5><ul>';
+					echo '<h5>' . $t->t('metaover_optional_not_found') . '</h5><ul>';
 					foreach ($entity['optional.notfound'] AS $key) {
 						echo '<li>' . htmlspecialchars($key) . '</li>';
 					}
@@ -84,7 +99,7 @@
 				
 				if (count($entity['leftovers']) > 0) {
 					echo '</div><div class="efieldlist warning">';
-					echo '<h5>The following fields was not reckognized</h5><ul>';
+					echo '<h5>' . $t->t('metaover_unknown_found') . '</h5><ul>';
 					foreach ($entity['leftovers'] AS $key => $value) {
 						echo '<li>' . htmlspecialchars($key) . '</li>';
 					}
@@ -96,28 +111,18 @@
 		}
 		
 		
-		if (array_key_exists('metadata.saml20-sp-hosted', $this->data)) 
-			showEntry('SAML 2.0 Service Provider (Hosted)', $this->data['metadata.saml20-sp-hosted'], $this->data['baseurlpath']);
-		if (array_key_exists('metadata.saml20-sp-remote', $this->data)) 
-			showEntry('SAML 2.0 Service Provider (Remote)', $this->data['metadata.saml20-sp-remote'], $this->data['baseurlpath']);
-		if (array_key_exists('metadata.saml20-idp-hosted', $this->data)) 
-			showEntry('SAML 2.0 Identity Provider (Hosted)', $this->data['metadata.saml20-idp-hosted'], $this->data['baseurlpath']);
-		if (array_key_exists('metadata.saml20-idp-remote', $this->data)) 
-			showEntry('SAML 2.0 Identity Provider (Remote)', $this->data['metadata.saml20-idp-remote'], $this->data['baseurlpath']);
+		showEntry($this, 'metadata.saml20-sp-hosted');
+		showEntry($this, 'metadata.saml20-sp-remote');
+		showEntry($this, 'metadata.saml20-idp-hosted');
+		showEntry($this, 'metadata.saml20-idp-remote');
 
-		if (array_key_exists('metadata.shib13-sp-hosted', $this->data)) 
-			showEntry('Shib 1.3 Service Provider (Hosted)', $this->data['metadata.shib13-sp-hosted'], $this->data['baseurlpath']);
-		if (array_key_exists('metadata.shib13-sp-remote', $this->data)) 
-			showEntry('Shib 1.3 Service Provider (Remote)', $this->data['metadata.shib13-sp-remote'], $this->data['baseurlpath']);
-		if (array_key_exists('metadata.shib13-idp-hosted', $this->data)) 
-			showEntry('Shib 1.3 Identity Provider (Hosted)', $this->data['metadata.shib13-idp-hosted'], $this->data['baseurlpath']);
-		if (array_key_exists('metadata.shib13-idp-remote', $this->data)) 
-			showEntry('Shib 1.3 Identity Provider (Remote)', $this->data['metadata.shib13-idp-remote'], $this->data['baseurlpath']);
+		showEntry($this, 'metadata.shib13-sp-hosted');
+		showEntry($this, 'metadata.shib13-sp-remote');
+		showEntry($this, 'metadata.shib13-idp-hosted');
+		showEntry($this, 'metadata.shib13-idp-remote');
 
-		if (array_key_exists('metadata.wsfed-sp-hosted', $this->data))
-			showEntry('WS-Federation Service Provider (Hosted)', $this->data['metadata.wsfed-sp-hosted'], $this->data['baseurlpath']);
-		if (array_key_exists('metadata.wsfed-idp-remote', $this->data))
-			showEntry('WS-Federation Identity Provider (Remote)', $this->data['metadata.wsfed-idp-remote'], $this->data['baseurlpath']);
+		showEntry($this, 'metadata.wsfed-sp-hosted');
+		showEntry($this, 'metadata.wsfed-idp-remote');
 
 		
 		?>
